requirements.txt
# Pour le module Telegram
python-telegram-bot==13.15
google-cloud-bigquery>=3.0.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
openai>=1.0.0
loguru>=0.7.0
